StatusCircle felleskomponent
