+++ src/apps/DisciplineReleaseControl/Components/Electro/ElectroView.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useQuery } from 'react-query';
+import { StatusCircle } from '../../../../packages/GardenUtils/src';
 import { EleNetwork } from '../../Types/eleNetwork';
 import { Pipetest } from '../../Types/pipetest';
-import { StatusCircle } from '../Garden/StatusCircle';
 import { ElectroNode } from './ElectroNode';
 import { getElectroViewCompletionStatusColor, getNodeStatus } from './electroViewHelpers';
 import { getEleNetworks } from './getEleNetworks';
